[7I] Split cli.py into cli_baseline, cli_config, cli_explain sub-modules
Extract three sub-Typer apps from cli.py (2328 LOC) into dedicated modules:
- cli_explain.py: explain_app Typer + 11 @explain_app.command handlers
- cli_config.py: config_app Typer + config init handler
- cli_baseline.py: baseline_app Typer + create/show/clear/update/diff
  handlers + _load_baseline_or_exit/_load_baseline_or_warn/
  _write_baseline_or_exit/_run_scan_for_baseline/_entry_count_label helpers

cli.py retains: scan, report, history, fix, dashboard, install-hook,
uninstall-hook, init, setup commands + all shared orchestration helpers.
Imports sub-app objects via standard top-of-file imports; no circular
imports (cli_baseline uses a lazy import for _load_scan_config only).

cli.py: 2328 → 1976 LOC. All CLI commands, help text, and exit codes
unchanged. 1729 tests pass.
